Use int64 instead of int32 for response sizes.
Previously, the reported size of responses larger than max int32 would
overflow, and these overflowed sizes were even surfaced to UI for the
Data Reduction Proxy.

Verified by building Chrome with and without this patch, and using the
Data Saver Chrome extension to download a 2.6 GB file. Without this
patch, a negative number is shown in the Data Saver stats. With this
patch, 2.6 GB is shown in the Data Saver stats.

BUG=475081

Review URL: https://codereview.chromium.org/1074293003

Cr-Commit-Position: refs/heads/master@{#324738}
(cherry picked from commit b5c82196f9bb0b0033a29ff0b14553e4c2e7f41d)

Review URL: https://codereview.chromium.org/1084683002

Cr-Commit-Position: refs/branch-heads/2357@{#72}
Cr-Branched-From: 59d4494849b405682265ed5d3f5164573b9a939b-refs/heads/master@{#323860}

+++ components/data_reduction_proxy/core/browser/data_reduction_proxy_metrics.cc
@@ -210,7 +210,7 @@ class DailyContentLengthUpdate {
   }
 
   // Update the lengths for the current day.
-  void Add(int content_length) {
+  void Add(int64 content_length) {
     AddInt64ToListPref(kNumDaysInHistory - 1, content_length, update_);
   }
 
@@ -275,7 +275,7 @@ class DailyDataSavingUpdate {
   }
 
   // Update the lengths for the current day.
-  void Add(int original_content_length, int received_content_length) {
+  void Add(int64 original_content_length, int64 received_content_length) {
     original_.Add(original_content_length);
     received_.Add(received_content_length);
   }
@@ -351,8 +351,8 @@ int64 GetAdjustedOriginalContentLength(
 }
 
 void UpdateContentLengthPrefsForDataReductionProxy(
-    int received_content_length,
-    int original_content_length,
+    int64 received_content_length,
+    int64 original_content_length,
     bool with_data_reduction_proxy_enabled,
     DataReductionProxyRequestType request_type,
     base::Time now,
